archive_to_paperless project version

+++ src/archive_to_paperless.cs
@@ -1,41 +1,39 @@
 #!/usr/local/share/dotnet/dotnet
+
+// to publish, use `cp archive_to_paperless.cs /Users/pmario/Local/tools/archive_to_paperless.cs`
+
 using System;
 using System.IO;
 
-string outputFile = System.IO.Path.GetTempPath() + "Schnellaktion_log.txt";
-
-void AppendLog(string message)
-{
-    message = $"{DateTime.Now.ToString("s")} - {message}";
-
-    System.Console.WriteLine(message);
-    using (var writer = new StreamWriter(outputFile, append: true))
-    {
-        writer.WriteLine(message);
-    }
-}
-System.Console.WriteLine($"Logging to {outputFile}");
-
 if (args.Length == 0)
 {
-    AppendLog("No arguments provided.\n");
+    Console.Error.WriteLine("No arguments provided.\n");
     return 1;
 }
 
-var archivePath = "/Users/pmario/Local/archived";
-var NASDropPath = "/Volumes/dropfolders/paperless";
+const string archivePath = "/Users/pmario/Local/archived";
+const string NASDropPath = "/Volumes/dropfolders/paperless";
+// const string NASDropMount = "//contrib@truenas._smb._tcp.local/dropfolders";
 
 if (!Directory.Exists(NASDropPath))
 {
-    AppendLog($"NAS drop path not found:  {NASDropPath}\n");
+    Console.Error.WriteLine($"NAS drop path not mounted / found:  {NASDropPath}\n");
     return 1;
 }
 
+int result = 0;
 foreach (var arg in args)
 {
+    string ext = Path.GetExtension(arg);
+    if (string.Compare(ext, ".pdf", StringComparison.OrdinalIgnoreCase) != 0)
+    {
+        Console.WriteLine($"Ignoring extension: {ext}");
+        continue;
+    }
+
     if (!File.Exists(arg))
     {
-        AppendLog($"File not found:  {arg}\n");
+        Console.Error.WriteLine($"File not found:  {arg}\n");
         continue;
     }
     try
@@ -43,18 +41,35 @@ foreach (var arg in args)
         var fileName = Path.GetFileName(arg);
         var destPath = Path.Combine(NASDropPath, fileName);
 
-        File.Copy(arg, destPath);
+        if (!File.Exists(destPath))
+        {
+            File.Copy(arg, destPath);
+        }
 
         var archiveDestPath = Path.Combine(archivePath, fileName);
-        File.Move(arg, archiveDestPath);        
+        if (File.Exists(archiveDestPath))
+        {
+            if (File.GetLastWriteTimeUtc(archiveDestPath) == File.GetLastWriteTimeUtc(arg))
+                Console.WriteLine($"File: {fileName}  already archived!");
+            else
+                Console.WriteLine($"File: {fileName}  already archived, but different write times!");
+        }
+        else
+        {
+            File.Move(arg, archiveDestPath);
+        }
     }
-    catch (System.Exception ex)
+    catch (Exception ex)
     {
-        var sb = new System.Text.StringBuilder();
-        sb.AppendLine($"Error processing file:  {arg}");
-        sb.AppendLine($"   {ex}");
+        string msg = $"Error processing file:  {arg}  / {ex.Message}";
+        Console.Error.WriteLine(msg);
 
-        AppendLog(sb.ToString());
+        var dest = Environment.GetFolderPath(Environment.SpecialFolder.Desktop);
+        dest = Path.Combine(dest, "script_error.txt");
+        File.AppendText(msg);
+        result = -1;
     }
 }
-return 0;
+
+Console.WriteLine("SUCCESS!");
+return result;
